requirements.txt
# Django
Django>=3.0.7

# Parser reqs
requests>=2.23.0
lxml>=4.5.1

# REST
djangorestframework>=3.11.0

# Celery reqs
celery>=4.4.5
redis>=3.5.3
celerybeat-redis>=0.1.5